changed the overlay for playlist creation to seperate pages #17
