+++ src/components/LoadingPlaylist.tsx
@@ -8,14 +8,14 @@ interface GenerateProps {
   artistID: string;
   danceMin: number;
   danceMax: number;
-  onClose: () => void;
+  // onClose: () => void;
 }
 
 const LoadingPlaylist: React.FC<GenerateProps> = ({
   artistID,
   danceMin,
   danceMax,
-  onClose,
+  // onClose,
 }) => {
   const tokenFromLocalStorage = localStorage.getItem("accessTokenLocal");
   const { setPlaylistData } = useContext(CollectContext);
@@ -39,12 +39,12 @@ const LoadingPlaylist: React.FC<GenerateProps> = ({
       navigate("/result");
 
       // Close the overlay, if the current URL path is '/result' insteat of reload
-      onClose();
+      // onClose();
       // Check if the current URL path is '/result'
-      if (window.location.pathname === "/result") {
-        // Reload the page if on the '/result' page
-        window.location.reload();
-      }
+      // if (window.location.pathname === "/result") {
+      //   // Reload the page if on the '/result' page
+      //   window.location.reload();
+      // }
     } catch (error) {
       if (axios.isAxiosError(error)) {
         console.log("Error searching for artists:", error.response?.data);
